+++ cpp/example_code/dynamodb/dynamodb_partiql_batch_scenario.cpp
@@ -15,23 +15,18 @@
  * Purpose
  *
  * Demonstrates using the AWS SDK for C++ to create an Amazon DynamoDB table and
- *  to perform a series of operations on the table using PartiQL.
+ *  to perform a series of operations on the table using PartiQL batch statements.
  *
  * 1. Create a table with partition: year, and sort: title. (CreateTable)
- * 2. Add a new movie using an "Insert" statement. (ExecuteStatement)
- * 3. Get the data for the movie using a "Select" statement. (ExecuteStatement)
- * 4. Update the data for the movie using an "Update" statement. (ExecuteStatement)
- * 5. Get the updated data for the movie using a "Select" statement. (ExecuteStatement)
- * 6. Delete the movie using a "Delete" statement. (ExecuteStatement)
- * 7. Add multiple movies using "Insert" statements. (BatchExecuteStatement)
- * 8. Get the data for multiple movies using "Select" statements. (BatchExecuteStatement)
- * 9. Update the data for multiple movies using "Update" statements. (BatchExecuteStatement)
- * 10. Get the updated data for multiple movies using "Select" statements. (BatchExecuteStatement)
- * 11. Delete multiple movies using "Delete" statements. (BatchExecuteStatement)
- * 12. Delete the table. (DeleteTable)
+ * 2. Add multiple movies using "Insert" statements. (BatchExecuteStatement)
+ * 3. Get the data for multiple movies using "Select" statements. (BatchExecuteStatement)
+ * 4. Update the data for multiple movies using "Update" statements. (BatchExecuteStatement)
+ * 5. Get the updated data for multiple movies using "Select" statements. (BatchExecuteStatement)
+ * 6. Delete multiple movies using "Delete" statements. (BatchExecuteStatement)
+ * 7. Delete the table. (DeleteTable)
  */
 
-#include "dyanamodb_samples.h"
+#include "dynamodb_samples.h"
 #include <iostream>
 #include <iomanip>
 #include <aws/core/Aws.h>
@@ -41,225 +36,6 @@
 
 #include <array>
 
-namespace AwsDoc {
-    namespace DynamoDB {
-
-        //! Scenario to modify and query a DynamoDB table using single and batch PartiQL
-        //! statements.
-        /*!
-          \sa dynamodbPartiqlScenario()
-          \param clientConfiguration: Aws client configuration.
-          \return bool: Function succeeded.
-         */
-        static bool dynamodbPartiqlScenario(
-                const Aws::Client::ClientConfiguration &clientConfiguration);
-
-    } //  namespace DynamoDB
-} // namespace AwsDoc
-
-bool
-
-// snippet-start:[cpp.example_code.dynamodb.Scenario_PartiQL_Single]
-//! Scenario to modify and query a DynamoDB table using single PartiQL statements.
-/*!
-  \sa partiqlExecuteScenario()
-  \param clientConfiguration: Aws client configuration.
-  \return bool: Function succeeded.
- */
-AwsDoc::DynamoDB::partiqlExecuteScenario(
-        const Aws::Client::ClientConfiguration &clientConfiguration) {
-    Aws::DynamoDB::DynamoDBClient dynamoClient(clientConfiguration);
-
-    // 2. Add a new movie using an "Insert" statement. (ExecuteStatement)
-    Aws::String title;
-    float rating;
-    int year;
-    Aws::String plot;
-    {
-        title = askQuestion(
-                "Enter the title of a movie you want to add to the table: ");
-        year = askQuestionForInt("What year was it released? ");
-        rating = askQuestionForFloatRange("On a scale of 1 - 10, how do you rate it? ",
-                                          1, 10);
-        plot = askQuestion("Summarize the plot for me: ");
-
-        Aws::DynamoDB::Model::ExecuteStatementRequest request;
-        std::stringstream sqlStream;
-        sqlStream << "INSERT INTO \"" << MOVIE_TABLE_NAME << "\" VALUE {'"
-                  << TITLE_KEY << "': ?, '" << YEAR_KEY << "': ?, '"
-                  << INFO_KEY << "': ?}";
-
-        request.SetStatement(sqlStream.str());
-
-        // Create the parameter attributes.
-        Aws::Vector<Aws::DynamoDB::Model::AttributeValue> attributes;
-        attributes.push_back(Aws::DynamoDB::Model::AttributeValue().SetS(title));
-        attributes.push_back(Aws::DynamoDB::Model::AttributeValue().SetN(year));
-
-        Aws::DynamoDB::Model::AttributeValue infoMapAttribute;
-
-        std::shared_ptr<Aws::DynamoDB::Model::AttributeValue> ratingAttribute = Aws::MakeShared<Aws::DynamoDB::Model::AttributeValue>(
-                ALLOCATION_TAG.c_str());
-        ratingAttribute->SetN(rating);
-        infoMapAttribute.AddMEntry(RATING_KEY, ratingAttribute);
-
-        std::shared_ptr<Aws::DynamoDB::Model::AttributeValue> plotAttibute = Aws::MakeShared<Aws::DynamoDB::Model::AttributeValue>(
-                ALLOCATION_TAG.c_str());
-        plotAttibute->SetS(plot);
-        infoMapAttribute.AddMEntry(PLOT_KEY, plotAttibute);
-        attributes.push_back(infoMapAttribute);
-        request.SetParameters(attributes);
-
-        Aws::DynamoDB::Model::ExecuteStatementOutcome outcome = dynamoClient.ExecuteStatement(
-                request);
-
-        if (!outcome.IsSuccess()) {
-            std::cerr << "Failed to add a movie: " << outcome.GetError().GetMessage()
-                      << std::endl;
-            return false;
-        }
-    }
-
-    std::cout << "\nAdded '" << title << "' to '" << MOVIE_TABLE_NAME << "'."
-              << std::endl;
-
-    //  3. Get the data for the movie using a "Select" statement. (ExecuteStatement)
-    {
-        Aws::DynamoDB::Model::ExecuteStatementRequest request;
-        std::stringstream sqlStream;
-        sqlStream << "SELECT * FROM  \"" << MOVIE_TABLE_NAME << "\" WHERE "
-                  << TITLE_KEY << "=? and " << YEAR_KEY << "=?";
-
-        request.SetStatement(sqlStream.str());
-
-        Aws::Vector<Aws::DynamoDB::Model::AttributeValue> attributes;
-        attributes.push_back(Aws::DynamoDB::Model::AttributeValue().SetS(title));
-        attributes.push_back(Aws::DynamoDB::Model::AttributeValue().SetN(year));
-        request.SetParameters(attributes);
-
-        Aws::DynamoDB::Model::ExecuteStatementOutcome outcome = dynamoClient.ExecuteStatement(
-                request);
-
-        if (!outcome.IsSuccess()) {
-            std::cerr << "Failed to retrieve movie information: "
-                      << outcome.GetError().GetMessage() << std::endl;
-            return false;
-        }
-        else {
-            // Print the retrieved movie information.
-            const Aws::DynamoDB::Model::ExecuteStatementResult &result = outcome.GetResult();
-
-            const Aws::Vector<Aws::Map<Aws::String, Aws::DynamoDB::Model::AttributeValue>> &items = result.GetItems();
-
-            if (items.size() == 1) {
-                printMovieInfo(items[0]);
-            }
-            else {
-                std::cerr << "Error: " << items.size() << " movies were retrieved. "
-                          << " There should be only one movie." << std::endl;
-            }
-        }
-    }
-
-    //  4. Update the data for the movie using an "Update" statement. (ExecuteStatement)
-    {
-        rating = askQuestionForFloatRange(
-                Aws::String("\nLet's update your movie.\nYou rated it  ") +
-                std::to_string(rating)
-                + ", what new rating would you give it? ", 1, 10);
-
-        Aws::DynamoDB::Model::ExecuteStatementRequest request;
-        std::stringstream sqlStream;
-        sqlStream << "UPDATE \"" << MOVIE_TABLE_NAME << "\" SET "
-                  << INFO_KEY << "." << RATING_KEY << "=? WHERE "
-                  << TITLE_KEY << "=? AND " << YEAR_KEY << "=?";
-
-        request.SetStatement(sqlStream.str());
-
-        Aws::Vector<Aws::DynamoDB::Model::AttributeValue> attributes;
-        attributes.push_back(Aws::DynamoDB::Model::AttributeValue().SetN(rating));
-        attributes.push_back(Aws::DynamoDB::Model::AttributeValue().SetS(title));
-        attributes.push_back(Aws::DynamoDB::Model::AttributeValue().SetN(year));
-
-        request.SetParameters(attributes);
-
-        Aws::DynamoDB::Model::ExecuteStatementOutcome outcome = dynamoClient.ExecuteStatement(
-                request);
-
-        if (!outcome.IsSuccess()) {
-            std::cerr << "Failed to update a movie: "
-                      << outcome.GetError().GetMessage();
-            return false;
-        }
-    }
-
-    std::cout << "\nUpdated '" << title << "' with new attributes:" << std::endl;
-
-    //  5. Get the updated data for the movie using a "Select" statement. (ExecuteStatement)
-    {
-        Aws::DynamoDB::Model::ExecuteStatementRequest request;
-        std::stringstream sqlStream;
-        sqlStream << "SELECT * FROM  \"" << MOVIE_TABLE_NAME << "\" WHERE "
-                  << TITLE_KEY << "=? and " << YEAR_KEY << "=?";
-
-        request.SetStatement(sqlStream.str());
-
-        Aws::Vector<Aws::DynamoDB::Model::AttributeValue> attributes;
-        attributes.push_back(Aws::DynamoDB::Model::AttributeValue().SetS(title));
-        attributes.push_back(Aws::DynamoDB::Model::AttributeValue().SetN(year));
-        request.SetParameters(attributes);
-
-        Aws::DynamoDB::Model::ExecuteStatementOutcome outcome = dynamoClient.ExecuteStatement(
-                request);
-        if (!outcome.IsSuccess()) {
-            std::cerr << "Failed to retrieve the movie information: "
-                      << outcome.GetError().GetMessage() << std::endl;
-            return false;
-        }
-        else {
-            const Aws::DynamoDB::Model::ExecuteStatementResult &result = outcome.GetResult();
-
-            const Aws::Vector<Aws::Map<Aws::String, Aws::DynamoDB::Model::AttributeValue>> &items = result.GetItems();
-
-            if (items.size() == 1) {
-                printMovieInfo(items[0]);
-            }
-            else {
-                std::cerr << "Error: " << items.size() << " movies were retrieved. "
-                          << " There should be only one movie." << std::endl;
-            }
-        }
-    }
-
-    std::cout << "Deleting the movie" << std::endl;
-
-    // Delete the movie using a "Delete" statement. (ExecuteStatement)
-    {
-        Aws::DynamoDB::Model::ExecuteStatementRequest request;
-        std::stringstream sqlStream;
-        sqlStream << "DELETE FROM  \"" << MOVIE_TABLE_NAME << "\" WHERE "
-                  << TITLE_KEY << "=? and " << YEAR_KEY << "=?";
-
-        request.SetStatement(sqlStream.str());
-
-        Aws::Vector<Aws::DynamoDB::Model::AttributeValue> attributes;
-        attributes.push_back(Aws::DynamoDB::Model::AttributeValue().SetS(title));
-        attributes.push_back(Aws::DynamoDB::Model::AttributeValue().SetN(year));
-        request.SetParameters(attributes);
-
-        Aws::DynamoDB::Model::ExecuteStatementOutcome outcome = dynamoClient.ExecuteStatement(
-                request);
-        if (!outcome.IsSuccess()) {
-            std::cerr << "Failed to delete the movie: "
-                      << outcome.GetError().GetMessage() << std::endl;
-            return false;
-        }
-    }
-
-    std::cout << "Movie successfully deleted." << std::endl;
-    return true;
-}
-// snippet-end:[cpp.example_code.dynamodb.Scenario_PartiQL_Single]
 
 // snippet-start:[cpp.example_code.dynamodb.Scenario_PartiQL_Batch]
 //! Scenario to modify and query a DynamoDB table using PartiQL batch statements.
@@ -270,8 +46,9 @@ AwsDoc::DynamoDB::partiqlExecuteScenario(
  */
 bool AwsDoc::DynamoDB::partiqlBatchExecuteScenario(
         const Aws::Client::ClientConfiguration &clientConfiguration) {
-    std::cout << "Now we will work with batches of movies" << std::endl;
 
+    // snippet-start:[cpp.example_code.dynamodb.BatchExecuteStatement.Insert]
+    // 2. Add multiple movies using "Insert" statements. (BatchExecuteStatement)
     Aws::DynamoDB::DynamoDBClient dynamoClient(clientConfiguration);
 
     std::vector<Aws::String> titles;
@@ -299,7 +76,6 @@ bool AwsDoc::DynamoDB::partiqlBatchExecuteScenario(
               << (titles.size() == 1 ? " movie " : " movies ")
               << "to the table using a batch \"INSERT\" statement." << std::endl;
 
-    // 7. Add multiple movies using "Insert" statements. (BatchExecuteStatement)
     {
         Aws::Vector<Aws::DynamoDB::Model::BatchStatementRequest> statements(
                 titles.size());
@@ -347,11 +123,13 @@ bool AwsDoc::DynamoDB::partiqlBatchExecuteScenario(
             return false;
         }
     }
+// snippet-end:[cpp.example_code.dynamodb.BatchExecuteStatement.Insert]
 
     std::cout << "Retrieving the movie data with a batch \"SELECT\" statement."
               << std::endl;
 
-    // 8. Get the data for multiple movies using "Select" statements. (BatchExecuteStatement)
+// snippet-start:[cpp.example_code.dynamodb.BatchExecuteStatement.Select]
+    // 3. Get the data for multiple movies using "Select" statements. (BatchExecuteStatement)
     {
         Aws::Vector<Aws::DynamoDB::Model::BatchStatementRequest> statements(
                 titles.size());
@@ -393,8 +171,10 @@ bool AwsDoc::DynamoDB::partiqlBatchExecuteScenario(
             return false;
         }
     }
+// snippet-end:[cpp.example_code.dynamodb.BatchExecuteStatement.Select]
 
-    // 9. Update the data for multiple movies using "Update" statements. (BatchExecuteStatement)
+// snippet-start:[cpp.example_code.dynamodb.BatchExecuteStatement.Update]
+    // 4. Update the data for multiple movies using "Update" statements. (BatchExecuteStatement)
 
     for (size_t i = 0; i < titles.size(); ++i) {
         ratings[i] = askQuestionForFloatRange(
@@ -403,7 +183,7 @@ bool AwsDoc::DynamoDB::partiqlBatchExecuteScenario(
                 + ", what new rating would you give it? ", 1, 10);
     }
 
-    std::cout << "Updating the movie with batch \"UPDATE\" statement." << std::endl;
+    std::cout << "Updating the movie with a batch \"UPDATE\" statement." << std::endl;
 
     {
         Aws::Vector<Aws::DynamoDB::Model::BatchStatementRequest> statements(
@@ -440,11 +220,12 @@ bool AwsDoc::DynamoDB::partiqlBatchExecuteScenario(
             return false;
         }
     }
+// snippet-end:[cpp.example_code.dynamodb.BatchExecuteStatement.Update]
 
     std::cout << "Retrieving the updated movie data with a batch \"SELECT\" statement."
               << std::endl;
 
-    // 10. Get the updated data for multiple movies using "Select" statements. (BatchExecuteStatement)
+    // 5. Get the updated data for multiple movies using "Select" statements. (BatchExecuteStatement)
     {
         Aws::Vector<Aws::DynamoDB::Model::BatchStatementRequest> statements(
                 titles.size());
@@ -486,12 +267,13 @@ bool AwsDoc::DynamoDB::partiqlBatchExecuteScenario(
             return false;
         }
     }
-
-    // 11. Delete multiple movies using "Delete" statements. (BatchExecuteStatement)
+// snippet-end:[cpp.example_code.dynamodb.BatchExecuteStatement.Update]
 
     std::cout << "Deleting the movie data with a batch \"DELETE\" statement."
               << std::endl;
 
+// snippet-start:[cpp.example_code.dynamodb.BatchExecuteStatement.Delete]
+    // 6. Delete multiple movies using "Delete" statements. (BatchExecuteStatement)
     {
         Aws::Vector<Aws::DynamoDB::Model::BatchStatementRequest> statements(
                 titles.size());
@@ -516,13 +298,14 @@ bool AwsDoc::DynamoDB::partiqlBatchExecuteScenario(
 
         Aws::DynamoDB::Model::BatchExecuteStatementOutcome outcome = dynamoClient.BatchExecuteStatement(
                 request);
-        
+
         if (!outcome.IsSuccess()) {
             std::cerr << "Failed to delete the movies: "
                       << outcome.GetError().GetMessage() << std::endl;
             return false;
         }
     }
+// snippet-end`:[cpp.example_code.dynamodb.BatchExecuteStatement.Delete]
 
     return true;
 }
@@ -531,33 +314,35 @@ bool AwsDoc::DynamoDB::partiqlBatchExecuteScenario(
 #ifndef TESTING_BUILD
 
 int main(int argc, char **argv) {
-    (void) argc; // suppress unused warning
-    (void) argv; // suppress unused warning
+    (void) argc; // Suppress unused warning.
+    (void) argv; // Suppress unused warning.
     Aws::SDKOptions options;
     InitAPI(options);
 
     {
+        std::cout << std::setfill('*')
+                  << std::setw(AwsDoc::DynamoDB::ASTERIX_FILL_WIDTH) << " "
+                  << std::endl;
+        std::cout
+                << "Welcome to the Amazon DynamoDB PartiQL batch statements demo."
+                << std::endl;
+        std::cout << std::setfill('*')
+                  << std::setw(AwsDoc::DynamoDB::ASTERIX_FILL_WIDTH) << " "
+                  << std::endl;
+
+// snippet-start:[cpp.example_code.dynamodb.Scenario_PartiQL_Batch.main]
         Aws::Client::ClientConfiguration clientConfig;
         //  1. Create a table. (CreateTable)
         if (AwsDoc::DynamoDB::createDynamoDBTable(AwsDoc::DynamoDB::MOVIE_TABLE_NAME,
                                                   clientConfig)) {
 
-            std::cout << std::setfill('*')
-                      << std::setw(AwsDoc::DynamoDB::ASTERIX_FILL_WIDTH) << " "
-                      << std::endl;
-            std::cout << "Welcome to the Amazon DynamoDB PartiQL demo." << std::endl;
-            std::cout << std::setfill('*')
-                      << std::setw(AwsDoc::DynamoDB::ASTERIX_FILL_WIDTH) << " "
-                      << std::endl;
-
-            AwsDoc::DynamoDB::partiqlExecuteScenario(clientConfig);
-
             AwsDoc::DynamoDB::partiqlBatchExecuteScenario(clientConfig);
 
-            // 9.Delete the table. (DeleteTable)
+            // 7. Delete the table. (DeleteTable)
             AwsDoc::DynamoDB::deleteDynamoTable(AwsDoc::DynamoDB::MOVIE_TABLE_NAME,
                                                 clientConfig);
         }
+// snippet-end:[cpp.example_code.dynamodb.Scenario_PartiQL_Batch.main]
     }
 
     return 0;
