+++ cpp/example_code/dynamodb/dynamodb_partiql_single_scenario.cpp
@@ -1,0 +1,291 @@
+/*
+   Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+   SPDX-License-Identifier: Apache-2.0
+*/
+/**
+ * Before running this C++ code example, set up your development environment, including your credentials.
+ *
+ * For more information, see the following documentation topic:
+ *
+ * https://docs.aws.amazon.com/sdk-for-cpp/v1/developer-guide/getting-started.html
+ *
+ * For information on the structure of the code examples and how to build and run the examples, see
+ * https://docs.aws.amazon.com/sdk-for-cpp/v1/developer-guide/getting-started-code-examples.html.
+ *
+ * Purpose
+ *
+ * Demonstrates using the AWS SDK for C++ to create an Amazon DynamoDB table and
+ *  to perform a series of operations on the table using single PartiQL statements.
+ *
+ * 1. Create a table with partition: year, and sort: title. (CreateTable)
+ * 2. Add a new movie using an "Insert" statement. (ExecuteStatement)
+ * 3. Get the data for the movie using a "Select" statement. (ExecuteStatement)
+ * 4. Update the data for the movie using an "Update" statement. (ExecuteStatement)
+ * 5. Get the updated data for the movie using a "Select" statement. (ExecuteStatement)
+ * 6. Delete the movie using a "Delete" statement. (ExecuteStatement)
+ * 7. Delete the table. (DeleteTable)
+ */
+
+#include "dynamodb_samples.h"
+#include <iostream>
+#include <iomanip>
+#include <aws/core/Aws.h>
+#include <aws/dynamodb/DynamoDBClient.h>
+#include <aws/dynamodb/model/ExecuteStatementRequest.h>
+#include <aws/dynamodb/model/BatchExecuteStatementRequest.h>
+
+#include <array>
+
+// snippet-start:[cpp.example_code.dynamodb.Scenario_PartiQL_Single]
+//! Scenario to modify and query a DynamoDB table using single PartiQL statements.
+/*!
+  \sa partiqlExecuteScenario()
+  \param clientConfiguration: Aws client configuration.
+  \return bool: Function succeeded.
+ */
+bool
+AwsDoc::DynamoDB::partiqlExecuteScenario(
+        const Aws::Client::ClientConfiguration &clientConfiguration) {
+    // snippet-start:[cpp.example_code.dynamodb.ExecuteStatement.Insert]
+    Aws::DynamoDB::DynamoDBClient dynamoClient(clientConfiguration);
+
+    // 2. Add a new movie using an "Insert" statement. (ExecuteStatement)
+    Aws::String title;
+    float rating;
+    int year;
+    Aws::String plot;
+    {
+        title = askQuestion(
+                "Enter the title of a movie you want to add to the table: ");
+        year = askQuestionForInt("What year was it released? ");
+        rating = askQuestionForFloatRange("On a scale of 1 - 10, how do you rate it? ",
+                                          1, 10);
+        plot = askQuestion("Summarize the plot for me: ");
+
+        Aws::DynamoDB::Model::ExecuteStatementRequest request;
+        std::stringstream sqlStream;
+        sqlStream << "INSERT INTO \"" << MOVIE_TABLE_NAME << "\" VALUE {'"
+                  << TITLE_KEY << "': ?, '" << YEAR_KEY << "': ?, '"
+                  << INFO_KEY << "': ?}";
+
+        request.SetStatement(sqlStream.str());
+
+        // Create the parameter attributes.
+        Aws::Vector<Aws::DynamoDB::Model::AttributeValue> attributes;
+        attributes.push_back(Aws::DynamoDB::Model::AttributeValue().SetS(title));
+        attributes.push_back(Aws::DynamoDB::Model::AttributeValue().SetN(year));
+
+        Aws::DynamoDB::Model::AttributeValue infoMapAttribute;
+
+        std::shared_ptr<Aws::DynamoDB::Model::AttributeValue> ratingAttribute = Aws::MakeShared<Aws::DynamoDB::Model::AttributeValue>(
+                ALLOCATION_TAG.c_str());
+        ratingAttribute->SetN(rating);
+        infoMapAttribute.AddMEntry(RATING_KEY, ratingAttribute);
+
+        std::shared_ptr<Aws::DynamoDB::Model::AttributeValue> plotAttibute = Aws::MakeShared<Aws::DynamoDB::Model::AttributeValue>(
+                ALLOCATION_TAG.c_str());
+        plotAttibute->SetS(plot);
+        infoMapAttribute.AddMEntry(PLOT_KEY, plotAttibute);
+        attributes.push_back(infoMapAttribute);
+        request.SetParameters(attributes);
+
+        Aws::DynamoDB::Model::ExecuteStatementOutcome outcome = dynamoClient.ExecuteStatement(
+                request);
+
+        if (!outcome.IsSuccess()) {
+            std::cerr << "Failed to add a movie: " << outcome.GetError().GetMessage()
+                      << std::endl;
+            return false;
+        }
+    }
+    // snippet-end:[cpp.example_code.dynamodb.ExecuteStatement.Insert]
+
+    std::cout << "\nAdded '" << title << "' to '" << MOVIE_TABLE_NAME << "'."
+              << std::endl;
+
+    // snippet-start:[cpp.example_code.dynamodb.ExecuteStatement.Select]
+    //  3. Get the data for the movie using a "Select" statement. (ExecuteStatement)
+    {
+        Aws::DynamoDB::Model::ExecuteStatementRequest request;
+        std::stringstream sqlStream;
+        sqlStream << "SELECT * FROM  \"" << MOVIE_TABLE_NAME << "\" WHERE "
+                  << TITLE_KEY << "=? and " << YEAR_KEY << "=?";
+
+        request.SetStatement(sqlStream.str());
+
+        Aws::Vector<Aws::DynamoDB::Model::AttributeValue> attributes;
+        attributes.push_back(Aws::DynamoDB::Model::AttributeValue().SetS(title));
+        attributes.push_back(Aws::DynamoDB::Model::AttributeValue().SetN(year));
+        request.SetParameters(attributes);
+
+        Aws::DynamoDB::Model::ExecuteStatementOutcome outcome = dynamoClient.ExecuteStatement(
+                request);
+
+        if (!outcome.IsSuccess()) {
+            std::cerr << "Failed to retrieve movie information: "
+                      << outcome.GetError().GetMessage() << std::endl;
+            return false;
+        }
+        else {
+            // Print the retrieved movie information.
+            const Aws::DynamoDB::Model::ExecuteStatementResult &result = outcome.GetResult();
+
+            const Aws::Vector<Aws::Map<Aws::String, Aws::DynamoDB::Model::AttributeValue>> &items = result.GetItems();
+
+            if (items.size() == 1) {
+                printMovieInfo(items[0]);
+            }
+            else {
+                std::cerr << "Error: " << items.size() << " movies were retrieved. "
+                          << " There should be only one movie." << std::endl;
+            }
+        }
+    }
+    // snippet-end:[cpp.example_code.dynamodb.ExecuteStatement.Select]
+
+    // snippet-start:[cpp.example_code.dynamodb.ExecuteStatement.Update]
+    //  4. Update the data for the movie using an "Update" statement. (ExecuteStatement)
+    {
+        rating = askQuestionForFloatRange(
+                Aws::String("\nLet's update your movie.\nYou rated it  ") +
+                std::to_string(rating)
+                + ", what new rating would you give it? ", 1, 10);
+
+        Aws::DynamoDB::Model::ExecuteStatementRequest request;
+        std::stringstream sqlStream;
+        sqlStream << "UPDATE \"" << MOVIE_TABLE_NAME << "\" SET "
+                  << INFO_KEY << "." << RATING_KEY << "=? WHERE "
+                  << TITLE_KEY << "=? AND " << YEAR_KEY << "=?";
+
+        request.SetStatement(sqlStream.str());
+
+        Aws::Vector<Aws::DynamoDB::Model::AttributeValue> attributes;
+        attributes.push_back(Aws::DynamoDB::Model::AttributeValue().SetN(rating));
+        attributes.push_back(Aws::DynamoDB::Model::AttributeValue().SetS(title));
+        attributes.push_back(Aws::DynamoDB::Model::AttributeValue().SetN(year));
+
+        request.SetParameters(attributes);
+
+        Aws::DynamoDB::Model::ExecuteStatementOutcome outcome = dynamoClient.ExecuteStatement(
+                request);
+
+        if (!outcome.IsSuccess()) {
+            std::cerr << "Failed to update a movie: "
+                      << outcome.GetError().GetMessage();
+            return false;
+        }
+    }
+    // snippet-end:[cpp.example_code.dynamodb.ExecuteStatement.Update]
+
+    std::cout << "\nUpdated '" << title << "' with new attributes:" << std::endl;
+
+    //  5. Get the updated data for the movie using a "Select" statement. (ExecuteStatement)
+    {
+        Aws::DynamoDB::Model::ExecuteStatementRequest request;
+        std::stringstream sqlStream;
+        sqlStream << "SELECT * FROM  \"" << MOVIE_TABLE_NAME << "\" WHERE "
+                  << TITLE_KEY << "=? and " << YEAR_KEY << "=?";
+
+        request.SetStatement(sqlStream.str());
+
+        Aws::Vector<Aws::DynamoDB::Model::AttributeValue> attributes;
+        attributes.push_back(Aws::DynamoDB::Model::AttributeValue().SetS(title));
+        attributes.push_back(Aws::DynamoDB::Model::AttributeValue().SetN(year));
+        request.SetParameters(attributes);
+
+        Aws::DynamoDB::Model::ExecuteStatementOutcome outcome = dynamoClient.ExecuteStatement(
+                request);
+        if (!outcome.IsSuccess()) {
+            std::cerr << "Failed to retrieve the movie information: "
+                      << outcome.GetError().GetMessage() << std::endl;
+            return false;
+        }
+        else {
+            const Aws::DynamoDB::Model::ExecuteStatementResult &result = outcome.GetResult();
+
+            const Aws::Vector<Aws::Map<Aws::String, Aws::DynamoDB::Model::AttributeValue>> &items = result.GetItems();
+
+            if (items.size() == 1) {
+                printMovieInfo(items[0]);
+            }
+            else {
+                std::cerr << "Error: " << items.size() << " movies were retrieved. "
+                          << " There should be only one movie." << std::endl;
+            }
+        }
+    }
+
+    std::cout << "Deleting the movie" << std::endl;
+
+    // snippet-start:[cpp.example_code.dynamodb.ExecuteStatement.Delete]
+    // 6. Delete the movie using a "Delete" statement. (ExecuteStatement)
+    {
+        Aws::DynamoDB::Model::ExecuteStatementRequest request;
+        std::stringstream sqlStream;
+        sqlStream << "DELETE FROM  \"" << MOVIE_TABLE_NAME << "\" WHERE "
+                  << TITLE_KEY << "=? and " << YEAR_KEY << "=?";
+
+        request.SetStatement(sqlStream.str());
+
+        Aws::Vector<Aws::DynamoDB::Model::AttributeValue> attributes;
+        attributes.push_back(Aws::DynamoDB::Model::AttributeValue().SetS(title));
+        attributes.push_back(Aws::DynamoDB::Model::AttributeValue().SetN(year));
+        request.SetParameters(attributes);
+
+        Aws::DynamoDB::Model::ExecuteStatementOutcome outcome = dynamoClient.ExecuteStatement(
+                request);
+        if (!outcome.IsSuccess()) {
+            std::cerr << "Failed to delete the movie: "
+                      << outcome.GetError().GetMessage() << std::endl;
+            return false;
+        }
+    }
+    // snippet-end:[cpp.example_code.dynamodb.ExecuteStatement.Delete]
+
+    std::cout << "Movie successfully deleted." << std::endl;
+    return true;
+}
+// snippet-end:[cpp.example_code.dynamodb.Scenario_PartiQL_Single]
+
+
+#ifndef TESTING_BUILD
+
+int main(int argc, char **argv) {
+    (void) argc; // Suppress unused warning.
+    (void) argv; // Suppress unused warning.
+    Aws::SDKOptions options;
+    InitAPI(options);
+
+    {
+        std::cout << std::setfill('*')
+                  << std::setw(AwsDoc::DynamoDB::ASTERIX_FILL_WIDTH) << " "
+                  << std::endl;
+        std::cout
+                << "Welcome to the Amazon DynamoDB PartiQL single statements demo."
+                << std::endl;
+        std::cout << std::setfill('*')
+                  << std::setw(AwsDoc::DynamoDB::ASTERIX_FILL_WIDTH) << " "
+                  << std::endl;
+
+        Aws::Client::ClientConfiguration clientConfig;
+
+// snippet-start:[cpp.example_code.dynamodb.Scenario_PartiQL_Single.main]
+        //  1. Create a table. (CreateTable)
+        if (AwsDoc::DynamoDB::createDynamoDBTable(AwsDoc::DynamoDB::MOVIE_TABLE_NAME,
+                                                  clientConfig)) {
+
+            AwsDoc::DynamoDB::partiqlExecuteScenario(clientConfig);
+
+            // 7. Delete the table. (DeleteTable)
+            AwsDoc::DynamoDB::deleteDynamoTable(AwsDoc::DynamoDB::MOVIE_TABLE_NAME,
+                                                clientConfig);
+        }
+// snippet-end:[cpp.example_code.dynamodb.Scenario_PartiQL_Single.main]
+    }
+
+    return 0;
+}
+
+#endif // TESTING_BUILD
+
+
