+++ cpp/example_code/dynamodb/tests/gtest_dynamodb_getting_started_scenario.cpp
@@ -6,7 +6,7 @@
 #include <gtest/gtest.h>
 #include <fstream>
 #include "dynamodb_gtests.h"
-#include "dyanamodb_samples.h"
+#include "dynamodb_samples.h"
 
 namespace AwsDocTest {
     // NOLINTNEXTLINE (readability-named-parameter)
